+++ src/commands/stats.ts
@@ -1,5 +1,4 @@
 import { I_LIKE_JS } from "../constants.ts";
-import { CheckForPath, ParsePath } from "../functions/filesystem.ts";
 import { ColorString, LogStuff } from "../functions/io.ts";
 import { GetProjectEnvironment, SpotProject } from "../functions/projects.ts";
 import { NameProject } from "../functions/projects.ts";
@@ -7,11 +6,6 @@ import type { DenoPkgJson, NodePkgJson } from "../types/runtimes.ts";
 
 export default async function TheStatistics(target: string) {
     const project = await SpotProject(target);
-    if (!project || !(await CheckForPath(project))) {
-        await LogStuff(`We couldn't find a project in ${await ParsePath(target)}. What's up?`, "warn");
-        return;
-    }
-
     const env = await GetProjectEnvironment(project);
     const name = await NameProject(project, "all");